test8/requirements.txt
# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Embedding Libraries
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Vector Storage
faiss-cpu>=1.7.4

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
Pillow>=10.0.0
pytesseract>=0.3.10
python-magic>=0.4.27

# Web Frameworks
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0

# HTTP Requests
requests>=2.31.0
httpx>=0.24.0

# CLI
click>=8.1.0
rich>=13.0.0

# Configuration
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
# sqlite3 is built into Python, no installation needed

# Logging and Monitoring
structlog>=23.1.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For testing FastAPI

# Optional: Microsoft Graph API (if using real Microsoft RAG)
# azure-ai-textanalytics>=5.3.0
# azure-identity>=1.13.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0